ACQ-1535: adapt markup to new styles. relocate components according to new section
