+++ examples/ft-ui/__test__/anonymous-user.test.js
@@ -7,23 +7,23 @@ describe('examples/ft-ui', () => {
   describe('Header link elements', () => {
     it('renders the expected anonymous user Header link elements', async () => {
       await expect(page).toMatchElement('a[href="/myft"]', { text: 'myFT' })
-      await expect(page).toMatchElement('.o-header__nav-list--right a[href="/login?location=/"]', {
+      await expect(page).toMatchElement('.o-header__top-column--right a[href="/login?location=/"]', {
         text: 'Sign In'
       })
-      await expect(page).toMatchElement('.o-header__nav-list--right a[href^="/products?"]', {
+      await expect(page).toMatchElement('.o-header__top-column--right a[href^="/products?"]', {
         text: 'Subscribe'
       })
     })
 
     it('does not render the logged-in user Header link elements', async () => {
       await expect(page).not.toMatchElement(
-        '.o-header__nav-list--right a[href="https://www.ft.com/myaccount"]',
+        '.o-header__top-column--right a[href="https://www.ft.com/myaccount"]',
         {
           text: 'Settings & Account'
         }
       )
       await expect(page).not.toMatchElement(
-        `.o-header__nav-list--right a[href="https://markets.ft.com/data/portfolio/dashboard"]`,
+        `.o-header__top-column--right a[href="https://markets.ft.com/data/portfolio/dashboard"]`,
         { text: 'Portfolio' }
       )
     })
