+++ packages/dotcom-ui-header/src/components/top/partials.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { THeaderProps } from '../../interfaces'
 import BrandFtMastheadSvg from '../svg-components/BrandFtMasthead'
-import { NavListRightAnon, SubscribeButton } from '../navigation/partials'
+import { TNavMenuItem } from '@financial-times/dotcom-types-navigation'
 
 const HeaderWrapper = (props) => (
   <header
@@ -93,25 +93,46 @@ const TopColumnRightLoggedIn = (props: THeaderProps) => {
   const subscribeAction = props.data['navbar-right-anon']?.items?.[1]
   return (
     <div className="o-header__top-column o-header__top-column--right">
-      <ul className="o-header__nav-list o-header__nav-list--right" data-trackable="user-nav">
-        <li className="o-header__nav-item">
-          {!props.userIsSubscribed && subscribeAction && (
-            <SubscribeButton item={subscribeAction} variant={props.variant} />
-          )}
-        </li>
-        <li className="o-header__nav-item o-header__nav-item--hide-s">
-          <MyFt />
-        </li>
-      </ul>
+      {!props.userIsSubscribed && subscribeAction && (
+        <SubscribeButton item={subscribeAction} variant={props.variant} />
+      )}
+      <MyFt />
     </div>
   )
 }
 
-const TopColumnRightAnon = (props: THeaderProps) => {
-  const userNavAnonItems = props.data['navbar-right-anon'].items
+const SignInLink = ({ item, variant }: { item: TNavMenuItem; variant?: string }) => {
+  const setTabIndex = variant === 'sticky' ? { tabIndex: -1 } : null
+  return (
+    <a className="o-header__nav-link" href={item.url} data-trackable={item.label} {...setTabIndex}>
+      {item.label}
+    </a>
+  )
+}
+const SubscribeButton = ({ item, variant }: { item: TNavMenuItem; variant?: string }) => {
+  const setTabIndex = variant === 'sticky' ? { tabIndex: -1 } : null
+  return (
+    <a
+      className="o-header__top-button"
+      // Added as the result of a DAC audit. This will be confusing for users of voice activation software
+      // as it looks like a button but behaves like a link without this role.
+      role="button"
+      href={item.url}
+      data-trackable={item.label}
+      {...setTabIndex}
+    >
+      {item.label}
+    </a>
+  )
+}
+
+const TopColumnRightAnon = ({ items, variant }: { items: TNavMenuItem[]; variant?: string }) => {
+  // If user is anonymous the second list item is styled as a button
+  const [signInAction, subscribeAction] = items
   return (
     <div className="o-header__top-column o-header__top-column--right">
-      <NavListRightAnon items={userNavAnonItems} variant={props.variant} />
+      {subscribeAction && <SubscribeButton item={subscribeAction} variant={variant} />}
+      {signInAction && <SignInLink item={signInAction} variant={variant} />}
     </div>
   )
 }
@@ -120,8 +141,18 @@ const TopColumnRight = (props: THeaderProps) => {
   if (props.userIsLoggedIn) {
     return <TopColumnRightLoggedIn {...props} />
   } else {
-    return <TopColumnRightAnon {...props} />
+    const userNavAnonItems = props.data['navbar-right-anon'].items
+    return <TopColumnRightAnon items={userNavAnonItems} variant={props.variant} />
   }
 }
 
-export { HeaderWrapper, TopWrapper, TopColumnLeft, TopColumnCenter, TopColumnCenterNoLink, TopColumnRight }
+export {
+  HeaderWrapper,
+  TopWrapper,
+  TopColumnLeft,
+  TopColumnCenter,
+  TopColumnCenterNoLink,
+  TopColumnRight,
+  TopColumnRightAnon,
+  SubscribeButton
+}
